% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_shff.r
\name{is_shff}
\alias{is_shff}
\title{Check if an HDF5 dataset is shff-valid.}
\usage{
is_shff(h5in, dataset = NULL)
}
\arguments{
\item{h5in}{Input file.}

\item{dataset}{Dataset in input file to read or \code{NULL}. In the latter case (e.g. \code{NULL}), the dataset will be contained
within a group named as the input dataset}
}
\value{
Returns \code{TRUE} on success, \code{FALSE} otherwise.
}
\description{
Check if an HDF5 dataset is shff-valid.
}
