% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shff.r
\name{read_shff}
\alias{read_shff}
\title{read_h5df}
\usage{
read_shff(h5in, dataset = NULL, indices = NULL)
}
\arguments{
\item{h5in}{Input file.}

\item{dataset}{Dataset in input file to read or \code{NULL}. In the latter case, TODO}

\item{indices}{TODO}
}
\value{
A dataframe.
}
\description{
TODO
}
\examples{
library(hdfio)

f = system.file("exampledata/pytables_table.h5", package="hdfio")
df = read_h5df(f, "mydata")
df

}
\seealso{
\code{\link{write_h5df}}
}
