% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_shff.r
\name{write_shff}
\alias{write_shff}
\title{write_shff}
\usage{
write_shff(x, file, dataset = NULL, compression = 4)
}
\arguments{
\item{x}{Input dataset (dataframe or datatable)}

\item{file}{Output file.}

\item{dataset}{Dataset in input file to read or \code{NULL}. In the latter case (e.g. \code{NULL}), the dataset will be contained 
within a group named as the input dataset.}

\item{compression}{HDF5 compression level. An integer, 0 (least compression) to 9 (most compression).  Default is
\code{compression} = 4.}
}
\description{
write_shff
}
\details{
TODO
}
